Hotel regulations

+++ mainapp/models.py
@@ -68,7 +68,7 @@ class Regulation(db.Model):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String(100), nullable=False, unique=True)
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
 
     def __str__(self):
         return str(self.name)
